# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Contract Management System API
  description: |
    Complete API documentation for the Contract Management System.
    
    ## Authentication
    All API requests require authentication using JWT tokens.
    Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    - Anonymous requests: 20/hour
    - Authenticated requests: 1000/hour
    - Premium users: 5000/hour
    
    ## WebSocket Support
    Real-time updates are available via WebSocket connections at:
    - Production: wss://api.contractmanagement.com/ws
    - Development: ws://localhost:8000/ws
    
  version: 1.0.0
  contact:
    name: API Support
    email: api@contractmanagement.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.contractmanagement.com/v1
    description: Production server
  - url: https://staging-api.contractmanagement.com/v1
    description: Staging server
  - url: http://localhost:8000/api/v1
    description: Development server
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Template:
      type: object
      required:
        - name
        - category
        - content
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [employment, service, nda, sales, lease, legal, other]
        content:
          type: string
        variables:
          type: array
          items:
            $ref: '#/components/schemas/TemplateVariable'
        isPublic:
          type: boolean
          default: false
        price:
          type: number
          format: double
          minimum: 0
        rating:
          type: number
          format: float
          minimum: 0
          maximum: 5
          readOnly: true
        usageCount:
          type: integer
          readOnly: true
        tags:
          type: array
          items:
            type: string
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
    TemplateVariable:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [text, number, date, boolean, select]
        required:
          type: boolean
          default: false
        defaultValue:
          type: string
        options:
          type: array
          items:
            type: string
...
paths:
...
  /templates:
    get:
      tags:
        - Templates
      summary: List all templates
      operationId: listTemplates
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [employment, service, nda, sales, lease, legal, other]
        - name: isPublic
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
      responses:
        200:
          description: Templates retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  templates:
                    type: array
                    items:
                      $ref: '#/components/schemas/Template'
                      
    post:
      tags:
        - Templates
      summary: Create a new template
      operationId: createTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Template'
      responses:
        201:
          description: Template created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
  /templates/{id}:
    get:
      tags:
        - Templates
      summary: Get template by ID
      operationId: getTemplate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Template retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
  /marketplace/templates:
    get:
      tags:
        - Marketplace
      summary: Browse marketplace templates
      operationId: browseMarketplace
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: minRating
          in: query
          schema:
            type: number
            minimum: 0
            maximum: 5
      responses:
        200:
          description: Marketplace templates
          content:
            application/json:
              schema:
                type: object
                properties:
                  templates:
                    type: array
                    items:
                      $ref: '#/components/schemas/Template'
...